// Laboratorio10-Routing/next-ssg-isr-app/src/types/rickandmorty.ts
export interface Info {
  count: number;
  pages: number;
  next: string | null;
  prev: string | null;
}

export interface Character {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: { name: string; url: string };
  location: { name: string; url: string };
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export interface ApiListResponse<T> {
  info: Info;
  results: T[];
}